chore(connector-spring-configuration): Refactor bean configuration so that all bean used both in inbound and inbound connector are in a generic configuration
