+++ connector-runtime/spring-boot-starter-camunda-connectors/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,6 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 io.camunda.connector.runtime.InboundConnectorsAutoConfiguration,\
 io.camunda.connector.runtime.OutboundConnectorsAutoConfiguration,\
+io.camunda.connector.runtime.ConnectorsAutoConfiguration,\
 io.camunda.connector.runtime.WebhookConnectorAutoConfiguration,\
 io.camunda.connector.runtime.ConnectorsObservabilityAutoConfiguration
