+++ connector-runtime/spring-boot-starter-camunda-connectors/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,4 +1,5 @@
 io.camunda.connector.runtime.OutboundConnectorsAutoConfiguration
 io.camunda.connector.runtime.InboundConnectorsAutoConfiguration
+io.camunda.connector.runtime.ConnectorsAutoConfiguration
 io.camunda.connector.runtime.WebhookConnectorAutoConfiguration
 io.camunda.connector.runtime.ConnectorsObservabilityAutoConfiguration
